Exclude ui from Sonar and move navigation out of ui package
